Allow Updating the display with keyboard.
Pressing ENTER will now update the display. I really wanted the display to be
updated every time the current item changes, but could not figure out for the
life of me how to make "currentChanged" work. This is still a lot better than
forcing the user to use the mouse.